Restrict source files to .elm
When creating the list of elm source files in `PackageInfo`, I make sure
to filter-out all non .elm files that may lie in the source directories
of the elm project (such as .js native modules). `PackageInfo` is meant
to be used with this crate, and the crate processes only elm files.