Add precise location tracking to the lexer
The lexer now returns a pair (Location,ElmToken) in its iterator, this
should progress toward a precise location tracking organization.

Note: currently the location tracking is not integrated with the parser,
there is no tests for it, nor is the location tracked.

Surprisingly, I didn't need to do a lot of adaptation in the parser code
to accept that new type shape. The most complicated was making
`filter_indent` generic over the wrapper time.